+++ lib/commands/server_config.mli
@@ -51,6 +51,7 @@ type imapConfig = {
   domain: string; (* email domain, by default host name *)
   maildir_parse: bool; (* parse message into MIME parts when in maildir storage format *)
   single_store: bool; (* single-store attachments in irmin and workdir format, default true *)
+  hybrid: bool; (* hybrid of irmin and workdir store (store should be set to irmin, default false *)
 }
 
 val default_config : imapConfig