+++ CHANGES.md
@@ -54,3 +54,4 @@
 * add attachments compression option
 * add option to store irmin/workdir MIME parts as single store for dedup or in one blob for fast access
 * support for tls 0.5.0
+* add hybrid irmin/workdir storage - messages are stored in workdir on FS, metadata is stored in Git